Find the dominant water types

Find the dominant waters type in a group of samples, if there are more than n samples. Match HCO3 and HCO3* types. Considered dominant types if in >p of samples.

Usage

dominant_water_types(d, n = 5, p = 0.75)

Arguments

d

Data frame with water_type column (e.g., output of rems_to_aquachem())

n

Numeric. Number of samples above which to remove outliers.

p

Numeric. Proportion of data required to assign a dominant water type.

Value

Data frame with added dominant column (TRUE or FALSE) if the water type is in the dominant set.